Adding transaction service and pricing service

+++ src/main/java/com/derekfleming/sportsbetchallenge/api/dto/TicketDto.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class TicketDto {
     private String ticketType;
     private Integer quantity;
-    private Float totalCost;
+    private Double totalCost;
 }
